fw_env: Fix environment size and CRC on 64-bit hosts.
- Reference fram u-boot ea19527c209b0d97b532a8dc5bf28f2cdd016888

+++ fw_env/fw_env.c
@@ -218,7 +218,7 @@ static int get_config (char *);
 #endif
 static inline ulong getenvsize (void)
 {
-	ulong rc = CONFIG_ENV_SIZE - sizeof (long);
+	ulong rc = CONFIG_ENV_SIZE - sizeof (uint32_t);
 
 	if (HaveRedundEnv)
 		rc -= sizeof (char);
